+++ prod/Watchdogd/json.c
@@ -144,14 +144,14 @@
 /******************************************************************************************************************************/
  gdouble Json_get_double ( JsonNode *query, gchar *chaine )
   { GValue valeur = G_VALUE_INIT;
-    gint retour;
+    gdouble retour;
     JsonObject *object = json_node_get_object (query);
     JsonNode *node = json_object_get_member ( object, chaine );
     if (!node) { return(0.0); }
     json_node_get_value ( node, &valeur );
          if ( G_VALUE_HOLDS_STRING (&valeur) ) { retour = atof(g_value_get_string (&valeur)); }
     else if ( G_VALUE_HOLDS_DOUBLE (&valeur) ) { retour = g_value_get_double(&valeur); }
-    else { printf("%s: Erreur getting '%s'\n", __func__, chaine ); retour = -1; }
+    else { printf("%s: Erreur getting '%s'\n", __func__, chaine ); retour = -1.0; }
     g_value_unset ( &valeur );
     return(retour);
   }
